refactor: export models
